services/web_dashboard/requirements.txt
# Web Dashboard Service Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Cache
redis[hiredis]>=4.6.0

# Message Queue
aio-pika>=9.3.0

# HTTP client for API proxy
httpx>=0.25.0

# Template engine (optional, for advanced HTML generation)
# jinja2>=3.1.0

# Static files
# aiofiles>=23.0.0

# Logging
loguru>=0.7.0

# Utilities
python-dateutil>=2.8.2

# Shared services (import from parent)
# -shared.models
# -shared.database
# -shared.utils

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
